Planteo inicial ws
