+++ src/services/WsGameHandler.js
@@ -1,7 +1,12 @@
-import { useEffect } from "react";
-
 // Esta función se encarga de manejar todos los mensajes que llegan por websocket
-const wsGameHandler = (ws, fetchTurnData, getWinner, fetchFigs, fetchMovs) => {
+const wsGameHandler = (
+  ws,
+  fetchTurnData,
+  getWinner,
+  fetchFigs,
+  fetchMovs,
+  fetchBoxCards
+) => {
   useEffect(() => {
     if (!ws) return;
 
@@ -16,9 +21,11 @@ const wsGameHandler = (ws, fetchTurnData, getWinner, fetchFigs, fetchMovs) => {
       } else if (message === "Hay Ganador") {
         console.log("Mensaje de ganador recibido");
         getWinner();
+      } else if (message === "Hay modificación de Tablero") {
+        fetchBoxCards();
       }
     };
-  }, [ws, fetchTurnData, getWinner, fetchFigs, fetchMovs]);
+  }, [ws, fetchTurnData, getWinner, fetchFigs, fetchMovs, fetchBoxCards]);
 };
 
 export default wsGameHandler;
